requirements.txt
ipython==8.16.1
pandas==2.1.1
torch-geometric==2.3.1
torchvision==0.16.0+3324288
dagshub==0.3.8.post2
